Move install task to its own file
