+++ index.js
@@ -17,35 +17,8 @@ import 'babel-polyfill';
 import gulp from 'gulp';
 import gulpSequence from 'gulp-sequence';
 import requireDir from 'require-dir';
-import readline from 'readline';
-import { exec } from 'child_process';
-
-const getTaskDeps = (taskConstrucs, config) => {
-  const deps = {};
-
-  for(let key in taskConstrucs) {
-    if(typeof taskConstrucs[key].getTaskDeps === 'function') {
-      Object.assign(deps, taskConstrucs[key].getTaskDeps(config));
-    }
-  }
-
-  return Object.values(deps).map((x) => x.split('/')[0]);
-};
-
-const passConfigRecursive = (taskConstrucs, config) => {
-  for(let key in taskConstrucs) {
-    const task = taskConstrucs[key].default || taskConstrucs[key];
-    if(typeof task === 'object') {
-      passConfigRecursive(task, config);
-    }
-    else if(typeof task === 'function') {
-      task(config);
-    }
-    else {
-      throw new Error('Can only pass config to a function, or an object containing functions');
-    }
-  }
-};
+import { passConfigToTasks, getTaskDeps } from './taskHelpers';
+import installTask from './wsTasks/install';
 
 export function start(toRun, config) {
   config = Object.assign({
@@ -57,46 +30,10 @@ export function start(toRun, config) {
 
   // Load all of the tasks that we might need to run
   const taskConstrucs = requireDir('./tasks', { recurse: true });
-  passConfigRecursive(taskConstrucs, config);
+  passConfigToTasks(taskConstrucs, config);
 
-  gulp.task('install', () => {
-    const deps = getTaskDeps(taskConstrucs, config);
-
-    if(!deps.length) {
-      console.log('No dependencies are needed!');
-      process.exit(0);
-    }
-
-    console.log('The following npm modules will be installed, and saved to your package.json');
-    deps.forEach((x) => console.log(x));
-
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout
-    });
-
-    rl.question('Proceed? (y/n): ', (answer) => {
-      if(answer[0].toLowerCase() === 'y') {
-        const inst = exec('npm install --save ' + deps.join(' '), (error) => {
-          if(error) {
-            console.error(error);
-            process.exit(1);
-          }
-          else {
-            console.log('Modules successfully installed');
-            process.exit(0);
-          }
-        });
-
-        inst.stdout.pipe(process.stdout);
-        inst.stderr.pipe(process.stderr);
-      }
-      else {
-        console.log('Install cancelled');
-        process.exit(0);
-      }
-    });
-  });
+  const deps = getTaskDeps(taskConstrucs, config);
+  installTask(deps);
 
   // `gulp.start()` forces our tasks in parallel, so we need a pseudo-task to
   // force sequential task running
